Completed HW-27 and refactored all test cases

+++ src/config/global.setup.ts
@@ -1,7 +1,11 @@
+// Description:
+// This file contains options for global setup in Playwright tests.
+
 import path from "path";
 import { rimraf } from "rimraf";
 
 export default function globalSetup() {
+    if (process.env.CI) return;
     const resultsPath = path.resolve(__dirname, "allure-results");
     rimraf.sync(resultsPath);
 }
